requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dateutil
dateparser
requests
sqlalchemy
psycopg2-binary
pydantic